SVY-20206
Getting closer with the logs that we keep getting. Next time we should
be able to see exactly what stack adds the problem listeners back to a
destroyed DAL.

Fixed the previous auto-correct fix as well, but we shouldn't even reach
that now, as it should reject the listener creation in
addDataLinkedProperty when already destroyed.