SVY-20206
Dispose the incorrect RelatedListener that is in an incorrect state even
if it's no longer in the DAL (for auto correct) + avoid calling .clear
on a Collections.emptyList() that is immutable; although that seems to
not give an exception - it's just a no-op.

I still don't get how it is possible according to the latest logs that
the DAL is destroyed and no longer references a RelatedListener, but
that listener's listeners still trigger.