SVY-20345 investigate/implement if we can use the servoytheme.less.properties in a form.css file
changed the default how to setup the solution (properties) less and how
to use the servoy theme with it.

Now a solution_properties.less file is generated besides the
solution.less
and that properties.less imports the theme.properties (but not the
servoy theme itself anymore)
that servoy theme.less is now imported by the solution.less
This way only properties can be imported in a form.less file (not the
whole theme css

changed the default behavior of the themeloader now without version or
with version of 2025 then the default is just given the ng2 variants.