Fixed SVY-20451: Some properties that register themselves to DAL give warning because they think they are equal but they are not
The equals and hashCode additions were meant as a performance
improvement, but as long as solutions don't do things like
elements.mylabel.text = "sameTextAsBefore %someDP%" (so using the same
value as before) on many components all the time, it should not matter
much.

The equals/hashcode lead otherwise to unexpected problems in code logic.
The alternative would be to make all those smart values aware of all the
context they need to use in equals/hashcode in order to make sure they
don't report as equal when they are not actually equal...