Fix for SVYX-1091: When running a jasper report this error occours
A foundset based Jasper report containing a relation based sub-report
would give that error message. The sub-report used a different thread,
and it triggered a calculation when getting data thus firing record
event listeners in the wrong thread.

However, because Jasper did block the event thread while the other
thread was running the sub-report, this did not result in any real
problems. It was just like running all in the event thread.